[refactor] Moved carbon intensity chart to the carbon intensity file.
